Few slight changes before major changes
Still not a fan of sharp turn and forward. Doing it twice or 3 times in a row looks bad, so now it can only do one of those actions, then must travel left or right. I still want to change forward to be a bit more random.
The display also lights up and says what action it's doing now.